+++ index.d.ts
@@ -688,6 +688,13 @@ declare namespace Enumerable {
          */
         intersect(second: Sequence<T>, comparer?: EqualityComparer<T> | true): IEnumerable<T>;
         /**
+         * Checks if that sequence is empty or not
+         * by using the length() method.
+         *
+         * @return {boolean} Is Empty or not.
+         */
+        isEmpty(): boolean;
+        /**
          * Correlates the elements of that sequence and another based on matching keys.
          *
          * @param {Sequence<TInner>} inner The other sequence.
@@ -1436,6 +1443,8 @@ declare namespace Enumerable {
          */
         protected intersectInner(second: Array<T>, comparer: EqualityComparer<T>): IterableIterator<T>;
         /** @inheritdoc */
+        isEmpty(): boolean;
+        /** @inheritdoc */
         join<TInner = T, TOuterKey = TInnerKey | T, TInnerKey = TOuterKey | TInner, TResult = JoinedItems<T, TInner>>(inner: Sequence<TInner>, outerKeySelector?: Selector<T, TOuterKey>, innerKeySelector?: Selector<TInner, TInnerKey>, resultSelector?: (outer: T, inner: TInner) => TResult, keyEqualityComparer?: EqualityComparer<TOuterKey, TInnerKey> | true): IEnumerable<TResult>;
         /**
          * @see join()
@@ -1792,6 +1801,14 @@ declare namespace Enumerable {
      */
     function isEnumerable<T = any>(val: any): val is IEnumerable<T>;
     /**
+     * Checks if a sequence is (null) or empty.
+     *
+     * @param {IEnumerable<T>} seq The sequence to check.
+     *
+     * @return {boolean} Is (null) or empty.
+     */
+    function isNullOrEmpty<T>(seq: IEnumerable<T>): seq is (null | IEnumerable<T>);
+    /**
      * Checks if a value can be used as enumerable (sequence).
      *
      * @param {any} val The value to check.
@@ -1799,6 +1816,22 @@ declare namespace Enumerable {
      * @return {boolean} Is sequence or not.
      */
     function isSequence<T = any>(val: any): val is Sequence<T>;
+    /**
+     * Checks if a sequence is (undefined) / (null) or empty.
+     *
+     * @param {IEnumerable<T>} seq The sequence to check.
+     *
+     * @return {boolean} Is (undefined), (null) or empty.
+     */
+    function isUndefinedNullOrEmpty<T>(seq: IEnumerable<T>): seq is (undefined | null | IEnumerable<T>);
+    /**
+     * Checks if a sequence is (undefined) or empty.
+     *
+     * @param {IEnumerable<T>} seq The sequence to check.
+     *
+     * @return {boolean} Is (undefined) or empty.
+     */
+    function isUndefinedOrEmpty<T>(seq: IEnumerable<T>): seq is (undefined | IEnumerable<T>);
     /**
      * Checks if a value represents the NOT_FOUND symbol.
      *