+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 3.8.0 (November 8th, 2017; methods)
 
 * added `append()` and `appendArray()` methods
+* added `isEmpty()` method
+* added `isNullOrEmpty()`, `isUndefinedNullOrEmpty()` and `isUndefinedOrEmpty()` functions
 * added `prepend()` and `prependArray()` methods
 
 ## 3.7.1 (November 7th, 2017; methods)